Moved camera-taking functionality to DrawFragment
CameraFragment removed
Functionality moved to DrawFragment
MainActivity no longer uses CameraFragment
Camera button in DrawFragment now calls takePhoto method
DrawFragment uses factory method
