+++ app/src/main/java/edu/cascadia/doodlebug/DrawFragment.java
@@ -1,8 +1,8 @@
 package edu.cascadia.doodlebug;
 
 import android.app.Activity;
+import android.content.Intent;
 import android.graphics.Bitmap;
-import android.net.Uri;
 import android.os.Bundle;
 import android.app.Fragment;
 import android.view.LayoutInflater;
@@ -10,14 +10,24 @@ import android.view.View;
 import android.view.ViewGroup;
 
 public class DrawFragment extends Fragment {
-    private OnFragmentInteractionListener mListener;
     private DrawingView mView;
+
+    private final static int CAMERA_PIC_REQUEST = 2;
+    private static String TAKE_PHOTO = "TAKE_PHOTO";
 
     public DrawFragment() {}
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    public static DrawFragment newInstance(Boolean takePhoto) {
+        DrawFragment fragment = new DrawFragment();
+        Bundle args = new Bundle();
+        args.putBoolean(TAKE_PHOTO, takePhoto);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    public void onActivityCreated(Bundle b) {
+        if (getArguments().getBoolean(TAKE_PHOTO))
+            takePhoto();
     }
 
     @Override
@@ -25,30 +35,27 @@ public class DrawFragment extends Fragment {
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
         View v = inflater.inflate(R.layout.fragment_draw, container, false);
+        v.findViewById(R.id.imageButttonTakePic).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                takePhoto();
+            }
+        });
+
         mView = (DrawingView) v.findViewById(R.id.drawingView);
         return v;
     }
 
-    public void setBackground(Bitmap bitmap) { mView.setBackground(bitmap); }
+    public void takePhoto() {
+        startActivityForResult(
+                new Intent(android.provider.MediaStore.ACTION_IMAGE_CAPTURE),
+                CAMERA_PIC_REQUEST);
+    }
 
     @Override
-    public void onAttach(Activity activity) {
-        super.onAttach(activity);
-        try {
-            mListener = (OnFragmentInteractionListener) activity;
-        } catch (ClassCastException e) {
-            throw new ClassCastException(activity.toString()
-                    + " must implement OnFragmentInteractionListener");
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == CAMERA_PIC_REQUEST && resultCode == Activity.RESULT_OK) {
+            mView.setBackground((Bitmap) data.getExtras().get("data"));
         }
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        mListener = null;
-    }
-
-    public interface OnFragmentInteractionListener {
-        void startCamera();
     }
 }
